refactor(duckdb): Extract and compile generator
